Excel_AI_Bot_With_Feedback_and_MachineLearning/requirements.txt
streamlit
pandas
openpyxl
duckdb
python-dotenv
openai
